Deploy scripts work with minikube / minishift (#29)
* add Openshift README, delete unused var defs in policies and app scripts.

* README tweaks, utils tweaked for mini* env detection logic.

* fixed is_mini logic. retested w/ minikube and minishift

* add reinstall options for _mini*-start.sh

* fix master ingress address for minikube

+++ kubernetes/_minikube-start.sh
@@ -1,13 +1,18 @@
 #!/bin/bash
-rm -rf ~/.kube ~/.minikube
 if [[ "$KUBERNETES_VERSION" == "" ]]; then
 	echo "source _minikube-boot.env first before running this script."
-	echo "Ignore errors re: not getting Kubernetes host (cuz Kubernetes is not running yet, right?)"
 	exit -1
 fi
+# if no existing VM, delete past login state and minikube resources 
+if [[ "$(minikube status | grep minikube | awk '{print $2}')" == "" ]]; then
+  minikube delete
+  rm -rf ~/.kube ~/.minikube
+fi
 minikube config set memory $MINIKUBE_VM_MEMORY
-minikube start --memory $MINIKUBE_VM_MEMORY --vm-driver virtualbox --kubernetes-version $KUBERNETES_VERSION
-kubectl taint nodes minikube node-role.kubernetes.io/master:NoSchedule- node/minikube untainted
+minikube start --memory $MINIKUBE_VM_MEMORY --vm-driver virtualbox --kubernetes-version $KUBERNETES_VERSION 
+#remove all taints from the minikube node so that pods will get scheduled
+sleep 5
+kubectl patch node minikube -p '{"spec":{"taints":[]}}'
 echo ""
 echo "IMPORTANT!  IMPORTANT!  IMPORTANT!  IMPORTANT!"
 echo "You need to source _minikube-boot.env again to reference docker daemon in Minikube..."
